HPA successfully scales group size

+++ src/objects/hpa_pod_group.rs
@@ -8,9 +8,9 @@ pub struct HPAPodGroup {
     pub min_size: u64,
     pub max_size: u64,
 
-    pub min_group_cpu_percent: u64,
-    pub min_group_memory_percent: u64,
+    pub min_group_cpu_percent: f64,
+    pub min_group_memory_percent: f64,
 
-    pub max_group_cpu_percent: u64,
-    pub max_group_memory_percent: u64,
+    pub max_group_cpu_percent: f64,
+    pub max_group_memory_percent: f64,
 }
